Removing ValueInfo class and moving TimestampedValue to samza-api util
